remove and move code
