+++ hyrisecockpit/frontend/src/components/details/TimeInterval.vue
@@ -6,13 +6,17 @@
 import {
   defineComponent,
   SetupContext,
-  onMounted,
   computed,
   Ref,
   ref,
 } from "@vue/composition-api";
 import { Metric } from "@/types/metrics";
-import { useUpdatingInterval } from "@/meta/components";
+import {
+  getMetricRequestTime,
+  getMetricChartConfiguration,
+  getMetricDataType,
+} from "@/meta/metrics";
+import { useFormatting } from "@/meta/formatting";
 
 interface Props {
   metric: Metric;
@@ -35,6 +39,34 @@ export default defineComponent({
     };
   },
 });
+
+export function useUpdatingInterval(
+  context: SetupContext,
+  metric: Metric
+): Ref<string> {
+  const timestamps = context.root.$metricController.timestamps[metric];
+  const { formatDateWithoutMilliSec, formatDateToHHMMSS } = useFormatting();
+  return computed(() => {
+    let currentTimeStamp = formatDateWithoutMilliSec(new Date());
+    let type = "";
+    const intervalTime = Math.max(
+      Math.floor(getMetricRequestTime(metric) / Math.pow(10, 3)),
+      context.root.$selectionController.selectedPrecision.value
+    );
+    if (timestamps.value.length > 0) {
+      currentTimeStamp = timestamps.value[timestamps.value.length - 1];
+    }
+    if (getMetricDataType(metric) === "interval") {
+      type =
+        intervalTime > 1
+          ? `Aggregation: ${intervalTime} seconds`
+          : "Aggregation: 1 second";
+    } else {
+      type = `Snapshot: ${formatDateToHHMMSS(currentTimeStamp)}`;
+    }
+    return type;
+  });
+}
 </script>
 <style scoped>
 .interval {
