--- hyrisecockpit/frontend/src/helpers/testData.ts
@@ -1,731 +1,0 @@
-export function useGeneratingTestData(): {
-  generateThroughputData: () => number;
-  generateStorageData: () => any;
-  generateCPUData: () => any;
-  generateAccessData: () => any;
-} {
-  function generateRandomNumber(min: number, range: number) {
-    return Math.floor(Math.random() * range + min);
-  }
-
-  function generateThroughputData(): number {
-    return generateRandomNumber(1, 1000);
-  }
-  function generateStorageData(): any {
-    const json = {
-      body: {
-        storage: {
-          //parent
-          citadelle: {
-            //table
-            customer: {
-              data: {
-                //attribute
-                c_acctbal: {
-                  data_type: "float",
-                  encoding: ["Dictionary"],
-                  size: 589828,
-                },
-                c_address: {
-                  data_type: "string",
-                  encoding: ["Dictionary"],
-                  size: 3300064,
-                },
-                c_comment: {
-                  data_type: "string",
-                  encoding: ["Dictionary"],
-                  size: 3299864,
-                },
-              },
-              number_columns: 3,
-              size: 7189756,
-            },
-
-            lineitem: {
-              data: {
-                l_comment: {
-                  data_type: "string",
-                  encoding: ["LZ4"],
-                  size: 29715756,
-                },
-                l_commitdate: {
-                  data_type: "string",
-                  encoding: ["Dictionary"],
-                  size: 5504031,
-                },
-                l_discount: {
-                  data_type: "float",
-                  encoding: ["Dictionary"],
-                  size: 3002911,
-                },
-              },
-              number_columns: 3,
-              size: 38222698,
-            },
-          },
-        },
-      },
-      header: { message: "OK", status: 200 },
-    };
-    return json;
-  }
-
-  function generateAccessData(): any {
-    return {
-      body: {
-        chunks_data: {
-          york: {
-            customer: {
-              c_acctbal: [0, 0],
-              c_address: [0, 0],
-              c_comment: [0, 0],
-              c_custkey: [0, 0],
-              c_mktsegment: [0, 1],
-              c_name: [0, 0],
-              c_nationkey: [31, 54],
-              c_phone: [0, 0],
-            },
-            lineitem: {
-              l_comment: [
-                0,
-                0,
-                73,
-                0,
-                7,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-              ],
-              l_commitdate: [
-                0,
-                0,
-                0,
-                95,
-                0,
-                0,
-                0,
-                0,
-                0,
-                43,
-                0,
-                0,
-                61,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                25,
-                0,
-                68,
-                0,
-                0,
-                0,
-                0,
-                0,
-              ],
-              l_discount: [
-                18,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                21,
-                0,
-                0,
-                0,
-                0,
-                66,
-                0,
-                0,
-                0,
-                0,
-                78,
-                32,
-                54,
-                0,
-                0,
-                14,
-                0,
-                62,
-                0,
-                0,
-                0,
-                0,
-                0,
-              ],
-              l_extendedprice: [
-                0,
-                1,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                22,
-                0,
-                98,
-                0,
-                0,
-                0,
-                3,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                92,
-                19,
-                0,
-              ],
-              l_linenumber: [
-                25,
-                0,
-                0,
-                0,
-                0,
-                64,
-                0,
-                0,
-                0,
-                86,
-                0,
-                0,
-                0,
-                0,
-                0,
-                67,
-                14,
-                0,
-                0,
-                6,
-                0,
-                97,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-              ],
-              l_linestatus: [
-                0,
-                36,
-                76,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                57,
-                0,
-                0,
-                88,
-                0,
-                0,
-                48,
-                0,
-                0,
-                0,
-                24,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                55,
-                0,
-              ],
-              l_orderkey: [
-                0,
-                95,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                26,
-                0,
-                0,
-                80,
-                0,
-                0,
-                0,
-                91,
-                0,
-                0,
-                0,
-                7,
-                0,
-                0,
-                0,
-                22,
-                0,
-              ],
-              l_partkey: [
-                0,
-                0,
-                0,
-                18,
-                16,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                53,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                39,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                88,
-                0,
-                26,
-              ],
-              l_quantity: [
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                86,
-                0,
-                0,
-                0,
-                4,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                70,
-                0,
-                0,
-                16,
-                0,
-                0,
-                0,
-                62,
-              ],
-              l_receiptdate: [
-                0,
-                65,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                82,
-                0,
-                70,
-                0,
-                4,
-                0,
-                0,
-                16,
-                72,
-                0,
-                0,
-                84,
-                0,
-                0,
-                98,
-                0,
-                0,
-                0,
-                73,
-                0,
-                0,
-                0,
-              ],
-              l_returnflag: [
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                67,
-                0,
-                85,
-                0,
-                22,
-                0,
-                0,
-                0,
-                29,
-                0,
-                0,
-                0,
-                0,
-                71,
-                95,
-                0,
-                0,
-                27,
-                0,
-                62,
-                0,
-                11,
-                0,
-              ],
-              l_shipdate: [
-                68,
-                0,
-                0,
-                0,
-                0,
-                0,
-                60,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                88,
-                0,
-                43,
-                0,
-                0,
-                0,
-                0,
-                0,
-                42,
-                0,
-                0,
-                0,
-                0,
-                0,
-                1,
-                0,
-                0,
-              ],
-              l_shipinstruct: [
-                0,
-                0,
-                73,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                16,
-                0,
-                10,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                46,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                8,
-                0,
-              ],
-              l_shipmode: [
-                0,
-                0,
-                51,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                57,
-                64,
-                0,
-                0,
-                55,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-              ],
-              l_suppkey: [
-                93,
-                0,
-                6,
-                0,
-                0,
-                0,
-                76,
-                0,
-                0,
-                0,
-                17,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                64,
-                0,
-                59,
-                0,
-                0,
-                0,
-                84,
-                0,
-                15,
-                0,
-                0,
-                0,
-                33,
-              ],
-              l_tax: [
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                17,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
-                61,
-                0,
-                18,
-                0,
-                25,
-                0,
-                0,
-                24,
-              ],
-            },
-            nation: {
-              n_comment: [0],
-              n_name: [0],
-              n_nationkey: [0],
-              n_regionkey: [0],
-            },
-            orders: {
-              o_clerk: [0, 0, 0, 0, 97, 0, 12, 0],
-              o_comment: [0, 0, 0, 0, 0, 22, 0, 0],
-              o_custkey: [0, 0, 0, 0, 0, 0, 60, 0],
-              o_orderdate: [0, 0, 0, 0, 0, 0, 0, 16],
-              o_orderkey: [0, 0, 0, 0, 0, 78, 8, 0],
-              o_orderpriority: [0, 70, 0, 0, 0, 0, 0, 0],
-              o_orderstatus: [0, 0, 0, 0, 83, 0, 0, 0],
-              o_shippriority: [0, 0, 35, 99, 34, 0, 0, 1],
-              o_totalprice: [0, 0, 0, 0, 0, 0, 0, 33],
-            },
-            part: {
-              p_brand: [0],
-              p_comment: [0],
-              p_container: [95],
-              p_mfgr: [0],
-              p_name: [0],
-              p_partkey: [0],
-              p_retailsize: [0],
-              p_size: [10],
-              p_type: [0],
-            },
-            partsupp: {
-              ps_availqty: [0, 0, 0, 0],
-              ps_comment: [53, 0, 0, 0],
-              ps_partkey: [0, 0, 0, 77],
-              ps_suppkey: [0, 0, 0, 0],
-              ps_supplycost: [0, 0, 0, 0],
-            },
-            region: { r_comment: [0], r_name: [60], r_regionkey: [76] },
-            supplier: {
-              s_acctbal: [0],
-              s_address: [98],
-              s_comment: [62],
-              s_name: [0],
-              s_nationkey: [0],
-              s_phone: [0],
-              s_suppkey: [0],
-            },
-          },
-        },
-      },
-      header: { message: "OK", status: 200 },
-    };
-  }
-
-  function generateCPUData(): any {
-    return {
-      system_data: {
-        citadelle: {
-          cpu: [
-            47.7,
-            55.6,
-            86.4,
-            12.6,
-            0.2,
-            10.3,
-            43.7,
-            12.1,
-            42.3,
-            54.5,
-            98.6,
-            31.6,
-            84.7,
-            32.6,
-            50.8,
-            17.0,
-          ],
-          memory: {
-            active: 3657117696,
-            available: 8467795968,
-            buffers: 169537536,
-            cached: 4237438976,
-            free: 5536755712,
-            inactive: 2687451136,
-            percent: 32.1,
-            shared: 1149366272,
-            total: 12469334016,
-            used: 2525601792,
-          },
-        },
-        york: {
-          cpu: [
-            44.0,
-            87.1,
-            42.4,
-            41.5,
-            6.3,
-            53.0,
-            82.0,
-            94.1,
-            77.3,
-            15.0,
-            2.8,
-            72.9,
-            48.1,
-            86.6,
-            36.8,
-            90.3,
-          ],
-          memory: {
-            active: 3657117696,
-            available: 8467795968,
-            buffers: 169537536,
-            cached: 4237438976,
-            free: 5536755712,
-            inactive: 2687451136,
-            percent: 32.1,
-            shared: 1149366272,
-            total: 12469334016,
-            used: 2525601792,
-          },
-        },
-      },
-    };
-  }
-
-  return {
-    generateThroughputData,
-    generateStorageData,
-    generateCPUData,
-    generateAccessData,
-  };
-}
