+++ hyrisecockpit/frontend/src/meta/components.ts
@@ -1,12 +1,7 @@
-import { SetupContext, Ref, computed, watch } from "@vue/composition-api";
+import { SetupContext, Ref, computed } from "@vue/composition-api";
 import { MetricProps, ComparisonMetricData, Metric } from "@/types/metrics";
 import { useDataEvents } from "@/meta/events";
-import {
-  getMetricRequestTime,
-  getMetricChartConfiguration,
-  getMetricDataType,
-} from "@/meta/metrics";
-import { useFormatting } from "@/meta/formatting";
+import { getMetricChartConfiguration } from "@/meta/metrics";
 
 export function useLineChartComponent(
   props: MetricProps,
@@ -46,32 +41,4 @@ export function useDatabaseFlex(
       return { flex: `1 0 ${100 / props.selectedDatabases.length - 1}%` };
     }),
   };
-}
-
-export function useUpdatingInterval(
-  context: SetupContext,
-  metric: Metric
-): Ref<string> {
-  const timestamps = context.root.$metricController.timestamps[metric];
-  const { formatDateWithoutMilliSec, formatDateToHHMMSS } = useFormatting();
-  return computed(() => {
-    let currentTimeStamp = formatDateWithoutMilliSec(new Date());
-    let type = "";
-    const intervalTime = Math.max(
-      Math.floor(getMetricRequestTime(metric) / Math.pow(10, 3)),
-      context.root.$selectionController.selectedPrecision.value
-    );
-    if (timestamps.value.length > 0) {
-      currentTimeStamp = timestamps.value[timestamps.value.length - 1];
-    }
-    if (getMetricDataType(metric) === "interval") {
-      type =
-        intervalTime > 1
-          ? `Aggregation: ${intervalTime} seconds`
-          : "Aggregation: 1 second";
-    } else {
-      type = `Snapshot: ${formatDateToHHMMSS(currentTimeStamp)}`;
-    }
-    return type;
-  });
 }
