+++ hyrisecockpit/frontend/src/meta/metrics.ts
@@ -191,22 +191,6 @@ const metricsChartConfiguration: Record<Metric, ChartConfiguration> = {
   },
 };
 
-const metricDescription: Partial<Record<Metric, string>> = {
-  access:
-    "Number of accesses  <br/> separated by chunk and column  <br/> of the selected table.",
-  cpu: "Current processor workload.",
-  executedQueryTypeProportion:
-    "Distribution of types of <br/>  the executed queries.",
-  generatedQueryTypeProportion:
-    "Distribution of types of <br/>  the generated queries.",
-  latency: "Average latency <br/> when processing queries from queue.",
-  queueLength:
-    "Length of the queue <br/> containing queries which <br/> have to be processed by workers ",
-  ram: "Current memory usage of <br/>  the database instance in percent",
-  storage: "Sizes of the tables and <br/> columns currently available.",
-  throughput: "Number of queries <br/> processed in the last second.",
-};
-
 const metricDetailsConfiguration: Partial<Record<
   Metric,
   MetricDetailsConfiguration
@@ -281,10 +265,6 @@ export function getMetricChartConfiguration(
   metric: Metric
 ): ChartConfiguration {
   return metricsChartConfiguration[metric];
-}
-
-export function getMetricDescription(metric: Metric): string {
-  return metricDescription[metric]!;
 }
 
 export function getMetricDetailsConfiguration(
