--- hyrisecockpit/frontend/src/meta/views.ts
@@ -1,3 +1,0 @@
-import { PageName } from "@/types/views";
-
-export const pages: PageName[] = ["comparison", "workload", "overview"];
