+++ hyrisecockpit/frontend/src/plugins/vueRouter.ts
@@ -1,6 +1,5 @@
 import Vue from "vue";
 import VueRouter from "vue-router";
-import Instance from "../views/Instance.vue";
 import InstanceComparison from "../views/InstanceComparison.vue";
 import Overview from "../views/Overview.vue";
 import Home from "../views/Home.vue";
@@ -10,7 +9,6 @@ Vue.use(VueRouter);
 
 const routes = [
   { name: "home", path: "/", component: Home },
-  { name: "database", path: "/database/:id", component: Instance },
   { name: "workload", path: "/workload", component: WorkloadMonitoring },
   {
     name: "comparison",
