+++ hyrisecockpit/frontend/src/utils/methods.ts
@@ -1,4 +1,7 @@
-export function equals(array1: string[], array2: string[]): boolean {
+/** 
+detects equality of two different arrays
+*/
+export function equals<T>(array1: T[], array2: T[]): boolean {
   array1.forEach((table1) =>
     array2.forEach((table2) => {
       if (table1 !== table2) {
